+++ CHANGES.rst
@@ -11,6 +11,10 @@ Version 1.1
 
 Unreleased
 
+* Add ``sort`` argument to :meth:`~Reader.get_entries`;
+  currently only the implicit `recent` sort order is available.
+  (:issue:`105`)
+
 
 Version 1.0.1
 -------------